Move defer from utils to promise-utils
