+++ packages/atomic/src/components/common/item-list/item-list-common.ts
@@ -6,7 +6,8 @@ import {
   type FocusTargetController,
   getFirstFocusableDescendant,
 } from '@/src/utils/accessibility-utils';
-import {defer, once} from '@/src/utils/utils';
+import {defer} from '@/src/utils/promise-utils';
+import {once} from '@/src/utils/utils';
 import {updateBreakpoints} from '../../../utils/replace-breakpoint-utils';
 import type {CommerceStore} from '../../commerce/atomic-commerce-interface/store';
 import type {CommerceRecommendationStore} from '../../commerce/atomic-commerce-recommendation-interface/store';
