+++ packages/atomic/src/components/common/item-list/stencil-item-list-common.tsx
@@ -3,7 +3,8 @@ import {
   FocusTargetController,
   getFirstFocusableDescendant,
 } from '../../../utils/stencil-accessibility-utils';
-import {defer, once} from '../../../utils/utils';
+import {defer} from '../../../utils/promise-utils';
+import {once} from '../../../utils/utils';
 import {CommerceStore} from '../../commerce/atomic-commerce-interface/store';
 import {CommerceRecommendationStore} from '../../commerce/atomic-commerce-recommendation-interface/store';
 import {InsightStore} from '../../insight/atomic-insight-interface/store';
